+++ iOS/G3MiOSSDK/Specific/Derived/Canvas_iOS.mm
@@ -124,7 +124,7 @@ void Canvas_iOS::_setLineMiterLimit(float limit) {
 
 void Canvas_iOS::_setLineDash(float lengths[],
                               int count,
-                              int phase) {
+                              float phase) {
   CGContextSetLineDash(_context,
                        phase,
                        lengths, count);
